refactor: extract JiraChangeLogEventDataSource base class to eliminate duplication
